Changed datatype to BigInteger to handle large integer counts

+++ SplitInversion/ArrayAndInversion.cs
@@ -1,6 +1,8 @@
 using System.Linq;
+using System.Numerics;
+
 public class ArrayAndInversion
 {
     public int[] Arr { get; set; }
-    public int Inversions { get; set; }
+    public BigInteger Inversions { get; set; }
 }
